server/pyproject.toml

[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["Jan Wadolowski <189539@edu.p.lodz.pl>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.92.0"
uvicorn = "^0.20.0"
pydantic = "^1.10.5"
pymongo = {extras = ["srv"], version = "^4.3.3"}
python-dotenv = "^0.21.1"
pydantic-mongo = "^1.0.1"
loguru = "^0.6.0"

[tool.poetry.group.dev.dependencies]
mockito = "^1.4.0"
mypy = "^1.0.1"
bandit = "^1.7.4"
black = "^23.1.0"
isort = "^5.12.0"
autoflake = "^2.0.1"
pytest = "^7.2.1"
pydocstyle = "^6.3.0"
httpx = "^0.23.3"
pytest-asyncio = "^0.20.3"
pytest-cov = "^4.0.0"
freezegun = "^1.2.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
py_version=310
src_paths = ["src", "tests"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.black]
line-length = 88
target-version = ['py310']
exclude = '''
(
  /(
      \.git
    | \.mypy_cache
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.autoflake]
recursive = true
remove-all-unused-imports = true
ignore-init-module-imports = true
expand-star-imports = true
in-place = true

[tool.pytest]
asyncio_mode = "auto"